feat: refactor invoice duplication step definition into a separate file
